add common action models to shared models folder
