+++ secureli/actions/scan.py
@@ -6,10 +6,8 @@ from typing import Optional
 
 from secureli.abstractions.echo import EchoAbstraction
 from secureli.actions.action import (
-    VerifyOutcome,
     Action,
     ActionDependencies,
-    VerifyResult,
 )
 from secureli.models.exit_codes import ExitCode
 from secureli.models.publish_results import PublishResultsOption
@@ -20,6 +18,7 @@ from secureli.services.scanner import (
     ScannerService,
 )
 from secureli.settings import Settings
+from secureli.shared.models.actions import VerifyOutcome, VerifyResult
 from secureli.utilities.usage_stats import post_log, convert_failures_to_failure_count
 
 ONE_WEEK_IN_SECONDS: int = 7 * 24 * 60 * 60
