+++ tests/actions/test_action.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 import pytest
 from secureli.abstractions.pre_commit import InstallResult
 
-from secureli.actions.action import Action, ActionDependencies, VerifyOutcome
+from secureli.actions.action import Action, ActionDependencies
 from secureli.consts.logging import TELEMETRY_DEFAULT_ENDPOINT
 from secureli.models.echo import Color
 from secureli.repositories.secureli_config import SecureliConfig, VerifyConfigOutcome
@@ -13,6 +13,7 @@ from secureli.services.language_support import LanguageMetadata
 from secureli.services.scanner import ScanResult, Failure
 from secureli.services.updater import UpdateResult
 from secureli.settings import Settings
+from secureli.shared.models.actions import VerifyOutcome
 
 test_folder_path = Path("does-not-matter")
